Tratando Repostas de Código HTTP

+++ TransactionAPI/Transaction.API/Controllers/UserController.cs
@@ -19,20 +19,29 @@ namespace Transaction.API.Controllers
         public async Task<IActionResult> CreateUser([FromBody] CreateNewUserRequest request)
         {
             var response = await _userService.CreateNewUserAsync(request.Name, request.Age);
-            return Ok(response);
+            return CreatedAtAction(nameof(GetUserById), new { id = response.Id }, response);
         }
 
         [HttpDelete("delete-user/{id}")]
         public async Task<IActionResult> DeleteUser(int id)
         {
+            var user = await _userService.GetUserByIdAsync(id);
+
+            if (user == null)
+                return NotFound($"User with ID {id} not found.");
+
             await _userService.DeleteUserById(id);
-            return Ok();
+            return NoContent();
         }
 
         [HttpGet("get-all-users")]
         public async Task<ActionResult> GetAllUsers()
         {
             var (users, summary) = await _userService.GetAllUsersAsync();
+
+            if (!users.Any())
+                return NoContent();
+
             return Ok(new { users, summary });
         }
 
@@ -40,6 +49,10 @@ namespace Transaction.API.Controllers
         public async Task<IActionResult> GetUserById(int id)
         {
             var response = await _userService.GetUserByIdAsync(id);
+
+            if (response == null)
+                return NotFound($"User with ID {id} not found.");
+
             return Ok(response);
         }
     }
